--- components/icons/ArrowCircleBrokenUpLeftIcon.tsx
@@ -1,19 +1,0 @@
-import * as React from "react";
-const ArrowCircleBrokenUpLeftIcon = (props: React.SVGProps<SVGSVGElement>) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width={23}
-    height={22}
-    fill="none"
-    {...props}
-  >
-    <path
-      stroke="currentColor"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      strokeWidth={2}
-      d="M14.55 20.66a9.996 9.996 0 0 1-9.661-2.589C.984 14.166.984 7.834 4.889 3.93 8.794.024 15.126.024 19.03 3.929a9.996 9.996 0 0 1 2.59 9.66M8.96 14V8m0 0h6m-6 0 9.999 10"
-    />
-  </svg>
-);
-export default ArrowCircleBrokenUpLeftIcon;
