--- components/icons/BookOpenIcon.tsx
@@ -1,19 +1,0 @@
-import * as React from "react";
-const BookOpenIcon = (props: React.SVGProps<SVGSVGElement>) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width={23}
-    height={20}
-    fill="none"
-    {...props}
-  >
-    <path
-      stroke="currentColor"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      strokeWidth={2}
-      d="m11.48 19-.1-.15c-.694-1.042-1.042-1.563-1.5-1.94a4 4 0 0 0-1.378-.737C7.934 16 7.308 16 6.055 16H4.68c-1.12 0-1.68 0-2.108-.218a2 2 0 0 1-.874-.874c-.218-.428-.218-.988-.218-2.108V4.2c0-1.12 0-1.68.218-2.108a2 2 0 0 1 .874-.874C3 1 3.56 1 4.68 1h.4c2.24 0 3.36 0 4.216.436a4 4 0 0 1 1.748 1.748c.436.856.436 1.976.436 4.216m0 11.6V7.4m0 11.6.1-.15c.695-1.042 1.043-1.563 1.501-1.94a3.999 3.999 0 0 1 1.378-.737C15.027 16 15.653 16 16.906 16h1.375c1.12 0 1.68 0 2.108-.218a2 2 0 0 0 .873-.874c.218-.428.218-.988.218-2.108V4.2c0-1.12 0-1.68-.218-2.108a2 2 0 0 0-.873-.874C19.96 1 19.4 1 18.28 1h-.4c-2.24 0-3.36 0-4.216.436a4 4 0 0 0-1.749 1.748c-.436.856-.436 1.976-.436 4.216"
-    />
-  </svg>
-);
-export default BookOpenIcon;
