+++ components/icons/CustomIcons.tsx
@@ -1,0 +1,356 @@
+import * as React from "react";
+export const ArrowCircleBrokenUpLeftIcon = (
+  props: React.SVGProps<SVGSVGElement>
+) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={23}
+    height={22}
+    fill="none"
+    {...props}
+  >
+    <path
+      stroke="currentColor"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth={2}
+      d="M14.55 20.66a9.996 9.996 0 0 1-9.661-2.589C.984 14.166.984 7.834 4.889 3.93 8.794.024 15.126.024 19.03 3.929a9.996 9.996 0 0 1 2.59 9.66M8.96 14V8m0 0h6m-6 0 9.999 10"
+    />
+  </svg>
+);
+
+export const ArrowNarrowLeft = (props: React.SVGProps<SVGSVGElement>) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={16}
+    height={12}
+    fill="none"
+    {...props}
+  >
+    <path
+      stroke="currentColor"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth={1.5}
+      d="M14.627 6H1.293m0 0 5 5m-5-5 5-5"
+    />
+  </svg>
+);
+
+export const BankNoteIcon = (props: React.SVGProps<SVGSVGElement>) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={23}
+    height={16}
+    fill="none"
+    {...props}
+  >
+    <path
+      stroke="currentColor"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth={2}
+      d="M13.48 5h-2.5a1.5 1.5 0 1 0 0 3h1a1.5 1.5 0 0 1 0 3h-2.5m2-7v1m0 6v1m6-4h.01M5.48 8h.01M1.48 4.2v7.6c0 1.12 0 1.68.218 2.108a2 2 0 0 0 .874.874C3 15 3.56 15 4.68 15h13.6c1.12 0 1.68 0 2.108-.218a2 2 0 0 0 .874-.874c.218-.428.218-.988.218-2.108V4.2c0-1.12 0-1.68-.218-2.108a2 2 0 0 0-.873-.874C19.96 1 19.4 1 18.28 1H4.68C3.56 1 3 1 2.571 1.218a2 2 0 0 0-.874.874C1.48 2.52 1.48 3.08 1.48 4.2ZM17.98 8a.5.5 0 1 1-1 0 .5.5 0 0 1 1 0Zm-12 0a.5.5 0 1 1-1 0 .5.5 0 0 1 1 0Z"
+    />
+  </svg>
+);
+
+export const BookOpenIcon = (props: React.SVGProps<SVGSVGElement>) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={23}
+    height={20}
+    fill="none"
+    {...props}
+  >
+    <path
+      stroke="currentColor"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth={2}
+      d="m11.48 19-.1-.15c-.694-1.042-1.042-1.563-1.5-1.94a4 4 0 0 0-1.378-.737C7.934 16 7.308 16 6.055 16H4.68c-1.12 0-1.68 0-2.108-.218a2 2 0 0 1-.874-.874c-.218-.428-.218-.988-.218-2.108V4.2c0-1.12 0-1.68.218-2.108a2 2 0 0 1 .874-.874C3 1 3.56 1 4.68 1h.4c2.24 0 3.36 0 4.216.436a4 4 0 0 1 1.748 1.748c.436.856.436 1.976.436 4.216m0 11.6V7.4m0 11.6.1-.15c.695-1.042 1.043-1.563 1.501-1.94a3.999 3.999 0 0 1 1.378-.737C15.027 16 15.653 16 16.906 16h1.375c1.12 0 1.68 0 2.108-.218a2 2 0 0 0 .873-.874c.218-.428.218-.988.218-2.108V4.2c0-1.12 0-1.68-.218-2.108a2 2 0 0 0-.873-.874C19.96 1 19.4 1 18.28 1h-.4c-2.24 0-3.36 0-4.216.436a4 4 0 0 0-1.749 1.748c-.436.856-.436 1.976-.436 4.216"
+    />
+  </svg>
+);
+
+export const CarIcon = (props: React.SVGProps<SVGSVGElement>) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={18}
+    height={18}
+    fill="none"
+    {...props}
+  >
+    <path
+      stroke="currentColor"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth={1.5}
+      d="M14.25 7.5V4.65c0-.84 0-1.26-.164-1.581a1.5 1.5 0 0 0-.655-.656c-.32-.163-.74-.163-1.581-.163h-5.7c-.84 0-1.26 0-1.581.163a1.5 1.5 0 0 0-.656.656c-.163.32-.163.74-.163 1.581V7.5m0-.75H1.5V6m12.75.75h2.25V6m-12 4.125h.008m8.992 0h.008M5.1 7.5h7.8c1.26 0 1.89 0 2.371.245.424.216.768.56.984.984.245.48.245 1.11.245 2.371v2.4c0 .699 0 1.048-.114 1.324a1.5 1.5 0 0 1-.812.812c-.276.114-.625.114-1.324.114h-.45c-.279 0-.418 0-.535-.018a1.5 1.5 0 0 1-1.247-1.247C12 14.368 12 14.229 12 13.95c0-.07 0-.104-.005-.134a.375.375 0 0 0-.311-.311c-.03-.005-.064-.005-.134-.005h-5.1c-.07 0-.105 0-.134.005a.375.375 0 0 0-.311.311C6 13.846 6 13.88 6 13.95c0 .279 0 .418-.018.535a1.5 1.5 0 0 1-1.247 1.247c-.117.018-.256.018-.535.018h-.45c-.699 0-1.048 0-1.324-.114a1.5 1.5 0 0 1-.812-.812C1.5 14.548 1.5 14.199 1.5 13.5v-2.4c0-1.26 0-1.89.245-2.371a2.25 2.25 0 0 1 .984-.984C3.209 7.5 3.839 7.5 5.1 7.5Zm-.225 2.625a.375.375 0 1 1-.75 0 .375.375 0 0 1 .75 0Zm9 0a.375.375 0 1 1-.75 0 .375.375 0 0 1 .75 0Z"
+    />
+  </svg>
+);
+
+export const EditIcon = (props: React.SVGProps<SVGSVGElement>) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={12}
+    height={12}
+    fill="none"
+    {...props}
+  >
+    <path
+      stroke="currentColor"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth={1.5}
+      d="M6 10h4.5M8.25 1.75a1.06 1.06 0 0 1 1.5 1.5L3.5 9.5l-2 .5.5-2 6.25-6.25Z"
+    />
+  </svg>
+);
+
+export const HomeLineIcon = (props: React.SVGProps<SVGSVGElement>) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={21}
+    height={21}
+    fill="none"
+    {...props}
+  >
+    <path
+      stroke="currentColor"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth={2}
+      d="M6.48 16h8M9.498 1.764 2.715 7.04c-.453.353-.68.53-.843.75a2 2 0 0 0-.318.65c-.074.265-.074.552-.074 1.126V16.8c0 1.12 0 1.68.218 2.108a2 2 0 0 0 .874.874C3 20 3.56 20 4.68 20h11.6c1.12 0 1.68 0 2.108-.218a2 2 0 0 0 .874-.874c.218-.428.218-.988.218-2.108V9.565c0-.574 0-.861-.074-1.126a2.002 2.002 0 0 0-.318-.65c-.163-.22-.39-.397-.843-.75l-6.783-5.275c-.351-.273-.527-.41-.72-.462a1 1 0 0 0-.523 0c-.194.052-.37.189-.721.462Z"
+    />
+  </svg>
+);
+
+export const HostAiIcon = (props: React.SVGProps<SVGSVGElement>) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={33}
+    height={32}
+    fill="none"
+    {...props}
+  >
+    <path
+      fill="#fff"
+      d="M31.118 12.264v8.516a4.027 4.027 0 0 1-4.02 4.02h-5.04c-.53 0-1.048.163-1.493.475l-6.314 4.455a2.987 2.987 0 0 1-3.107.212 2.985 2.985 0 0 1-1.616-2.655v-2.238a.25.25 0 0 0-.246-.247H5.865a4.027 4.027 0 0 1-4.02-4.02v-8.517a5.272 5.272 0 0 1 2.8-4.66L14.99 2.092a3.149 3.149 0 0 1 2.97 0l10.356 5.511a5.274 5.274 0 0 1 2.801 4.662v-.001Z"
+    />
+    <path
+      fill="#3B9EFF"
+      d="M12.072 31.905c-.524 0-1.05-.124-1.535-.375a3.32 3.32 0 0 1-1.8-2.958V26.08a.28.28 0 0 0-.275-.276H4.655a4.483 4.483 0 0 1-4.477-4.478V11.84a5.872 5.872 0 0 1 3.116-5.191L14.822.508a3.515 3.515 0 0 1 3.308 0l11.534 6.138a5.872 5.872 0 0 1 3.117 5.193v9.486a4.483 4.483 0 0 1-4.478 4.477H22.69c-.59 0-1.166.183-1.664.53l-7.032 4.96a3.34 3.34 0 0 1-1.925.613h.003Zm4.404-28.463a.164.164 0 0 0-.081.02L4.868 9.602a2.529 2.529 0 0 0-1.343 2.237v9.487c0 .624.507 1.13 1.13 1.13h3.807a3.627 3.627 0 0 1 3.623 3.624v2.492l1.157-.846 5.869-4.137a6.253 6.253 0 0 1 3.582-1.134h5.613c.624 0 1.131-.507 1.131-1.13v-9.486c0-.939-.515-1.796-1.345-2.238L16.558 3.46a.174.174 0 0 0-.082-.02Z"
+    />
+    <path
+      fill="#10275B"
+      d="M10.206 15.014c.889 0 1.61-1.014 1.61-2.263 0-1.25-.721-2.263-1.61-2.263-.89 0-1.61 1.013-1.61 2.263s.72 2.263 1.61 2.263ZM22.758 15.014c.89 0 1.61-1.014 1.61-2.263 0-1.25-.72-2.263-1.61-2.263-.889 0-1.61 1.013-1.61 2.263s.721 2.263 1.61 2.263ZM20.137 18.342a3.688 3.688 0 0 1-3.653 3.172 3.706 3.706 0 0 1-3.662-3.172H20.137Z"
+    />
+  </svg>
+);
+
+export const InboxIcon = (props: React.SVGProps<SVGSVGElement>) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={23}
+    height={18}
+    fill="none"
+    {...props}
+  >
+    <path
+      stroke="currentColor"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth={2}
+      d="M1.98 9h3.382c.686 0 1.312.387 1.618 1 .307.613.933 1 1.618 1h5.764a1.81 1.81 0 0 0 1.618-1c.307-.613.933-1 1.619-1h3.381M8.447 1h6.067c1.077 0 1.615 0 2.09.164a3 3 0 0 1 1.122.693c.359.352.6.833 1.081 1.796l2.167 4.334c.189.377.283.566.35.765.059.175.102.356.127.54.03.207.03.419.03.841V12.2c0 1.68 0 2.52-.328 3.162a3 3 0 0 1-1.31 1.311C19.2 17 18.36 17 16.68 17H6.28c-1.68 0-2.52 0-3.162-.327a3 3 0 0 1-1.31-1.311C1.48 14.72 1.48 13.88 1.48 12.2v-2.067c0-.422 0-.634.03-.84a3 3 0 0 1 .127-.541c.067-.199.161-.388.35-.766l2.167-4.333c.481-.963.722-1.444 1.081-1.796a3 3 0 0 1 1.121-.693C6.832 1 7.37 1 8.447 1Z"
+    />
+  </svg>
+);
+
+export const LayersThreeIcon = (props: React.SVGProps<SVGSVGElement>) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={23}
+    height={22}
+    fill="none"
+    {...props}
+  >
+    <path
+      stroke="currentColor"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth={2}
+      d="m1.48 11 9.643 4.821c.13.066.197.099.265.111.061.012.124.012.185 0 .069-.012.134-.045.265-.11L21.481 11m-20 5 9.642 4.821c.13.066.197.099.265.111.061.012.124.012.185 0 .069-.012.134-.045.265-.11L21.481 16m-20-10 9.642-4.821c.13-.066.197-.098.265-.111a.5.5 0 0 1 .185 0c.069.013.134.045.265.11L21.481 6l-9.643 4.821a1.086 1.086 0 0 1-.265.111.501.501 0 0 1-.185 0c-.069-.012-.134-.045-.265-.11L1.48 6Z"
+    />
+  </svg>
+);
+
+export const LayoutAltIcon = (props: React.SVGProps<SVGSVGElement>) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={21}
+    height={20}
+    fill="none"
+    {...props}
+  >
+    <path
+      stroke="currentColor"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth={2}
+      d="M15.98 15h-11m11-4h-11m-3.5-4h18M6.28 1h8.4c1.68 0 2.52 0 3.162.327a3 3 0 0 1 1.312 1.311c.326.642.326 1.482.326 3.162v8.4c0 1.68 0 2.52-.326 3.162a3 3 0 0 1-1.312 1.311c-.641.327-1.481.327-3.161.327h-8.4c-1.68 0-2.52 0-3.162-.327a3 3 0 0 1-1.312-1.311C1.48 16.72 1.48 15.88 1.48 14.2V5.8c0-1.68 0-2.52.327-3.162A3 3 0 0 1 3.12 1.327C3.76 1 4.6 1 6.28 1Z"
+    />
+  </svg>
+);
+
+export const SvgComponent = (props: React.SVGProps<SVGSVGElement>) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={23}
+    height={18}
+    fill="none"
+    {...props}
+  >
+    <path
+      stroke="currentColor"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth={2}
+      d="m1.96 4 8.165 5.715c.661.463.992.695 1.351.784a2 2 0 0 0 .968 0c.36-.09.69-.32 1.351-.784L21.96 4M6.76 17h10.4c1.68 0 2.52 0 3.162-.327a3 3 0 0 0 1.311-1.311c.327-.642.327-1.482.327-3.162V5.8c0-1.68 0-2.52-.327-3.162a3 3 0 0 0-1.311-1.311C19.68 1 18.84 1 17.16 1H6.76c-1.68 0-2.52 0-3.162.327a3 3 0 0 0-1.311 1.311C1.96 3.28 1.96 4.12 1.96 5.8v6.4c0 1.68 0 2.52.327 3.162a3 3 0 0 0 1.311 1.311C4.24 17 5.08 17 6.76 17Z"
+    />
+  </svg>
+);
+
+export const MountainIcon = (props: React.SVGProps<SVGSVGElement>) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={17}
+    height={16}
+    fill="none"
+    {...props}
+  >
+    <g clipPath="url(#a)">
+      <path
+        stroke="currentColor"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+        strokeWidth={1.5}
+        d="M6.293 2 8.96 7.333 12.293 4l3.334 10H2.293l4-12Z"
+      />
+    </g>
+    <defs>
+      <clipPath id="a">
+        <path fill="#fff" d="M.96 0h16v16h-16z" />
+      </clipPath>
+    </defs>
+  </svg>
+);
+
+export const SettingsIcon = (props: React.SVGProps<SVGSVGElement>) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={21}
+    height={22}
+    fill="none"
+    {...props}
+  >
+    <path
+      stroke="currentColor"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth={2}
+      d="m7.876 18.371.584 1.315a2.212 2.212 0 0 0 4.044 0l.585-1.315a2.426 2.426 0 0 1 2.47-1.423l1.43.152a2.212 2.212 0 0 0 2.022-3.502l-.847-1.164a2.428 2.428 0 0 1-.46-1.434c0-.513.163-1.014.465-1.429l.846-1.163a2.211 2.211 0 0 0-2.022-3.502l-1.43.152a2.433 2.433 0 0 1-1.47-.312 2.426 2.426 0 0 1-1-1.117l-.589-1.315a2.212 2.212 0 0 0-4.044 0L7.876 3.63c-.208.468-.558.86-1 1.117-.445.256-.96.365-1.47.312L3.97 4.906A2.212 2.212 0 0 0 1.95 8.408l.847 1.163a2.429 2.429 0 0 1 0 2.858l-.847 1.163a2.211 2.211 0 0 0 2.022 3.502l1.43-.152a2.434 2.434 0 0 1 2.475 1.43Z"
+    />
+    <path
+      stroke="currentColor"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth={2}
+      d="M10.48 14a3 3 0 1 0 0-6 3 3 0 0 0 0 6Z"
+    />
+  </svg>
+);
+
+export const TreesIcon = (props: React.SVGProps<SVGSVGElement>) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={17}
+    height={16}
+    fill="none"
+    {...props}
+  >
+    <path
+      stroke="currentColor"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth={1.5}
+      d="M7.627 6.667V6.8a2 2 0 0 1-.733 3.867h-2.6A2 2 0 0 1 3.627 6.8v-.133a2 2 0 0 1 4 0ZM5.626 10.667v4M9.626 12.667v2"
+    />
+    <path
+      stroke="currentColor"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth={1.5}
+      d="M8.96 12.667h5.533a.668.668 0 0 0 .467-1.134l-2-2.2h.2a.667.667 0 0 0 .467-1.133l-2-2.2h.133a.667.667 0 0 0 .533-1.133L9.627 2l-.934 1"
+    />
+  </svg>
+);
+
+export const TriangleIcon = (props: React.SVGProps<SVGSVGElement>) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={11}
+    height={10}
+    fill="none"
+    {...props}
+  >
+    <path
+      fill="currentColor"
+      d="m3.46 9.054 5.833-3.333a.833.833 0 0 0 0-1.45L3.46.937a.833.833 0 0 0-1.25.73v6.666a.833.833 0 0 0 1.25.721Z"
+    />
+  </svg>
+);
+
+export const MailIcon = (props: React.SVGProps<SVGSVGElement>) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={23}
+    height={18}
+    fill="none"
+    {...props}
+  >
+    <path
+      stroke="currentColor"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth={2}
+      d="m1.96 4 8.165 5.715c.661.463.992.695 1.351.784a2 2 0 0 0 .968 0c.36-.09.69-.32 1.351-.784L21.96 4M6.76 17h10.4c1.68 0 2.52 0 3.162-.327a3 3 0 0 0 1.311-1.311c.327-.642.327-1.482.327-3.162V5.8c0-1.68 0-2.52-.327-3.162a3 3 0 0 0-1.311-1.311C19.68 1 18.84 1 17.16 1H6.76c-1.68 0-2.52 0-3.162.327a3 3 0 0 0-1.311 1.311C1.96 3.28 1.96 4.12 1.96 5.8v6.4c0 1.68 0 2.52.327 3.162a3 3 0 0 0 1.311 1.311C4.24 17 5.08 17 6.76 17Z"
+    />
+  </svg>
+);
+
+const SnowflakeIcon = (props: React.SVGProps<SVGSVGElement>) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={18}
+    height={18}
+    fill="none"
+    {...props}
+  >
+    <g clipPath="url(#a)">
+      <path
+        stroke="currentColor"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+        strokeWidth={1.5}
+        d="M9 6v6m0-6V1.5M9 6 5.25 2.25M9 6l3.75-3.75M9 12v4.5M9 12l-3.75 3.75M9 12l3.75 3.75M12 9H6m6 0h4.5M12 9l3.75-3.75M12 9l3.75 3.75M6 9H1.5M6 9 2.25 5.25M6 9l-3.75 3.75"
+      />
+    </g>
+    <defs>
+      <clipPath id="a">
+        <path fill="#fff" d="M0 0h18v18H0z" />
+      </clipPath>
+    </defs>
+  </svg>
+);
