--- components/icons/HostAiIcon.tsx
@@ -1,24 +1,0 @@
-import * as React from "react";
-const HostAiIcon = (props: React.SVGProps<SVGSVGElement>) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width={33}
-    height={32}
-    fill="none"
-    {...props}
-  >
-    <path
-      fill="#fff"
-      d="M31.118 12.264v8.516a4.027 4.027 0 0 1-4.02 4.02h-5.04c-.53 0-1.048.163-1.493.475l-6.314 4.455a2.987 2.987 0 0 1-3.107.212 2.985 2.985 0 0 1-1.616-2.655v-2.238a.25.25 0 0 0-.246-.247H5.865a4.027 4.027 0 0 1-4.02-4.02v-8.517a5.272 5.272 0 0 1 2.8-4.66L14.99 2.092a3.149 3.149 0 0 1 2.97 0l10.356 5.511a5.274 5.274 0 0 1 2.801 4.662v-.001Z"
-    />
-    <path
-      fill="#3B9EFF"
-      d="M12.072 31.905c-.524 0-1.05-.124-1.535-.375a3.32 3.32 0 0 1-1.8-2.958V26.08a.28.28 0 0 0-.275-.276H4.655a4.483 4.483 0 0 1-4.477-4.478V11.84a5.872 5.872 0 0 1 3.116-5.191L14.822.508a3.515 3.515 0 0 1 3.308 0l11.534 6.138a5.872 5.872 0 0 1 3.117 5.193v9.486a4.483 4.483 0 0 1-4.478 4.477H22.69c-.59 0-1.166.183-1.664.53l-7.032 4.96a3.34 3.34 0 0 1-1.925.613h.003Zm4.404-28.463a.164.164 0 0 0-.081.02L4.868 9.602a2.529 2.529 0 0 0-1.343 2.237v9.487c0 .624.507 1.13 1.13 1.13h3.807a3.627 3.627 0 0 1 3.623 3.624v2.492l1.157-.846 5.869-4.137a6.253 6.253 0 0 1 3.582-1.134h5.613c.624 0 1.131-.507 1.131-1.13v-9.486c0-.939-.515-1.796-1.345-2.238L16.558 3.46a.174.174 0 0 0-.082-.02Z"
-    />
-    <path
-      fill="#10275B"
-      d="M10.206 15.014c.889 0 1.61-1.014 1.61-2.263 0-1.25-.721-2.263-1.61-2.263-.89 0-1.61 1.013-1.61 2.263s.72 2.263 1.61 2.263ZM22.758 15.014c.89 0 1.61-1.014 1.61-2.263 0-1.25-.72-2.263-1.61-2.263-.889 0-1.61 1.013-1.61 2.263s.721 2.263 1.61 2.263ZM20.137 18.342a3.688 3.688 0 0 1-3.653 3.172 3.706 3.706 0 0 1-3.662-3.172H20.137Z"
-    />
-  </svg>
-);
-export default HostAiIcon;
