--- components/icons/InboxIcon.tsx
@@ -1,19 +1,0 @@
-import * as React from "react";
-const InboxIcon = (props: React.SVGProps<SVGSVGElement>) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width={23}
-    height={18}
-    fill="none"
-    {...props}
-  >
-    <path
-      stroke="currentColor"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      strokeWidth={2}
-      d="M1.98 9h3.382c.686 0 1.312.387 1.618 1 .307.613.933 1 1.618 1h5.764a1.81 1.81 0 0 0 1.618-1c.307-.613.933-1 1.619-1h3.381M8.447 1h6.067c1.077 0 1.615 0 2.09.164a3 3 0 0 1 1.122.693c.359.352.6.833 1.081 1.796l2.167 4.334c.189.377.283.566.35.765.059.175.102.356.127.54.03.207.03.419.03.841V12.2c0 1.68 0 2.52-.328 3.162a3 3 0 0 1-1.31 1.311C19.2 17 18.36 17 16.68 17H6.28c-1.68 0-2.52 0-3.162-.327a3 3 0 0 1-1.31-1.311C1.48 14.72 1.48 13.88 1.48 12.2v-2.067c0-.422 0-.634.03-.84a3 3 0 0 1 .127-.541c.067-.199.161-.388.35-.766l2.167-4.333c.481-.963.722-1.444 1.081-1.796a3 3 0 0 1 1.121-.693C6.832 1 7.37 1 8.447 1Z"
-    />
-  </svg>
-);
-export default InboxIcon;
