--- components/icons/LayersThreeIcon.tsx
@@ -1,19 +1,0 @@
-import * as React from "react";
-const LayersThreeIcon = (props: React.SVGProps<SVGSVGElement>) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width={23}
-    height={22}
-    fill="none"
-    {...props}
-  >
-    <path
-      stroke="currentColor"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      strokeWidth={2}
-      d="m1.48 11 9.643 4.821c.13.066.197.099.265.111.061.012.124.012.185 0 .069-.012.134-.045.265-.11L21.481 11m-20 5 9.642 4.821c.13.066.197.099.265.111.061.012.124.012.185 0 .069-.012.134-.045.265-.11L21.481 16m-20-10 9.642-4.821c.13-.066.197-.098.265-.111a.5.5 0 0 1 .185 0c.069.013.134.045.265.11L21.481 6l-9.643 4.821a1.086 1.086 0 0 1-.265.111.501.501 0 0 1-.185 0c-.069-.012-.134-.045-.265-.11L1.48 6Z"
-    />
-  </svg>
-);
-export default LayersThreeIcon;
