--- components/icons/LayoutAltIcon.tsx
@@ -1,19 +1,0 @@
-import * as React from "react";
-const LayoutAltIcon = (props: React.SVGProps<SVGSVGElement>) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width={21}
-    height={20}
-    fill="none"
-    {...props}
-  >
-    <path
-      stroke="currentColor"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      strokeWidth={2}
-      d="M15.98 15h-11m11-4h-11m-3.5-4h18M6.28 1h8.4c1.68 0 2.52 0 3.162.327a3 3 0 0 1 1.312 1.311c.326.642.326 1.482.326 3.162v8.4c0 1.68 0 2.52-.326 3.162a3 3 0 0 1-1.312 1.311c-.641.327-1.481.327-3.161.327h-8.4c-1.68 0-2.52 0-3.162-.327a3 3 0 0 1-1.312-1.311C1.48 16.72 1.48 15.88 1.48 14.2V5.8c0-1.68 0-2.52.327-3.162A3 3 0 0 1 3.12 1.327C3.76 1 4.6 1 6.28 1Z"
-    />
-  </svg>
-);
-export default LayoutAltIcon;
